Complete 64 bit changes for window numbers for X windows
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/branches/gnustep_testplant_branch@40207 72102866-910b-0410-8b05-ffd578937521

+++ Headers/Additions/GNUstepGUI/GSDisplayServer.h
@@ -105,7 +105,7 @@ APPKIT_EXPORT NSString *GSScreenNumber;
 - (NSArray *) screenList;
 
 - (void *) serverDevice;
-- (void *) windowDevice: (int)win;
+- (void *) windowDevice: (NSInteger)win;
 
 - (void) beep;
